[MARKET-315] Prevent NPE when quitting with open SpoonBrowser tabs on Windows
On Windows, browser tabs are instances of `org.eclipse.swt.ole.win32.OleClientSite`
that manages embedded ActiveX controls.

When disposed (at least in SWT 4.3.2) `OnInPlaceDeactivate` is called and if the
tab has focus forces its move to the next focusable control (the next open tab).
Great when closing a tab!

Unfortunately when quiting the app it seems it can end up on an already disposed tab,
throwing a NullPointerException.

Moving in advance the focus to other non-embedded control prevents it.