[PRD-5662]  Adding support for java.sql.Types.OTHER to ThinUtil
The OTHER type is mapped to TYPE_NONE and treated effectively as an
unknown type.

This has a corresponding change in the data service client to handle
setObject with a null value (in which case type is unknown).  Using
sql type OTHER in such cases for consistency with several other db
drivers (mysql, oracle, postgres, etc.)

http://jira.pentaho.com/browse/PRD-5662