[PDI-15162] "Running in parallel", CSV input step misses some records. (#2428)
*  [PDI-15162]  "Running in parallel", CSV input step misses some records.
 - Making a decision whether not to skip a line after checking whether we are actually starting with a new line, that wasn't read by the previous step (see javadocs or description bellow for more details)
 - Tests for parallel work of CSVInputStep written.

More detailed:
We need to skip row only if a line, that we are currently on is read by the previous step partly.
In other words, we DON'T skip a line if we are just beginning to read it from the first symbol.
We have to do some work for this: read last byte from the previous step and make sure that it is a new line byte.
But it's not enough. There could be a situation, where new line is indicated by '\r\n' construction. And if we are
between this construction, we want to skip last '\n', and don't want to include it in our line.

So, we DON'T skip line only if the previous char is new line indicator AND we are not between '\r\n'.

* [CLEANUP] Checkstyle applied